+++ src/main/java/com/example/backend/jwt/JWTFilter.java
@@ -1,6 +1,6 @@
 package com.example.backend.jwt;
 
-import com.example.backend.service.CustomUserDetailsService;
+import com.example.backend.service.member.CustomUserDetailsService;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
