+++ src/main/java/com/example/backend/model/entity/member/Member.java
@@ -1,6 +1,5 @@
-package com.example.backend.model.entity;
+package com.example.backend.model.entity.member;
 
-import com.example.backend.model.entity.UserRole;
 import jakarta.persistence.*;
 import lombok.*;
