+++ src/main/java/com/example/backend/model/repository/member/MemberRepository.java
@@ -1,7 +1,7 @@
-package com.example.backend.repository;
+package com.example.backend.model.repository.member;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.example.backend.model.entity.Member;
+import com.example.backend.model.entity.member.Member;
 
 import java.util.Optional;
