+++ src/main/java/com/example/backend/model/repository/member/ProfileRepository.java
@@ -1,6 +1,6 @@
-package com.example.backend.repository;
+package com.example.backend.model.repository.member;
 
-import com.example.backend.model.entity.Profile;
+import com.example.backend.model.entity.member.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfileRepository extends JpaRepository<Profile, Long> {
