+++ src/main/java/com/example/backend/service/member/MemberService.java
@@ -1,10 +1,10 @@
-package com.example.backend.service;
+package com.example.backend.service.member;
 
-import com.example.backend.dto.TechStackDTO;
-import com.example.backend.model.entity.Member;
-import com.example.backend.model.entity.Profile;
-import com.example.backend.repository.MemberRepository;
-import com.example.backend.repository.ProfileRepository;
+import com.example.backend.dto.member.TechStackDTO;
+import com.example.backend.model.entity.member.Member;
+import com.example.backend.model.entity.member.Profile;
+import com.example.backend.model.repository.member.MemberRepository;
+import com.example.backend.model.repository.member.ProfileRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
